requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

markdown2[all]
numpy==2.1.1
requests==2.32.3
sentencepiece==0.1.99
tokenizers>=0.15.2
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
wandb
shortuuid
httpx==0.24.0
deepspeed==0.12.3
peft==0.4.0
albumentations
opencv-python==4.10.0.84
tiktoken==0.6.0
accelerate==0.28.0
transformers==4.37.2
bitsandbytes==0.41.0
scikit-learn==1.2.2
einops==0.6.1
einops-exts==0.0.4
timm==0.6.13
Pillow==10.4.0
Shapely==2.0.6
apex==0.9.10dev
beautifulsoup4==4.12.3
boto3==1.35.28
fairscale==0.4.13
filetype==1.2.0
fitz==0.0.1.dev2
ftfy==6.2.3
huggingface_hub==0.24.6
jieba==0.42.1
megfile==3.1.3
natsort==8.4.0
nltk==3.9.1
pandas==2.2.3
playwright==1.47.0
pydantic==2.9.2
pydantic_settings==2.5.2
pypdfium2==4.30.0
pytesseract==0.3.13
python-dotenv==1.0.1
skimage==0.0
smart_open==7.0.4
tqdm==4.66.5
verovio==4.3.1
